pyproject.toml
[project]
name = "ai-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
